+++ tests/Pest.php
@@ -1,10 +1,9 @@
 <?php
 
-use Mockery;
 use Tests\TestCase;
 
 uses(TestCase::class)->in('Feature');
 
 afterEach(function () {
-    Mockery::close();
+    \Mockery::close();
 });